requirements.txt
absl-py==1.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
libtorrent==2.0.7
python-qbittorrent==0.4.3
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
yapf==0.33.0
